fix: issues with file paths on Windows
Turns out `std::fs::canonicalize(path)` returns a special path with a
'//?/' prefix on Windows, which basically throws it into a compatability
mode where long paths are allowed and '/' for entering directories is
not. This caused huge issues where we were unable to read any
directories properly, namely `starpkg build` without a `-d` argument.

Additionally, package paths on all targets are now made relative to
the current working directory (not `-d`!). This resolves the above issue
and gives nicer messages when we need to show a path to the user (e.g.
when displaying an error or warning).